editor/libeditor/text/tests/test_bug692520.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=692520
-->
<head>
  <title>Test for Bug 692520</title>
  <script type="application/javascript" src="/MochiKit/packed.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=692520">Mozilla Bug 692520</a>
<p id="display"></p>
<div id="content">
<textarea></textarea>
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 692520 **/
function test(prop, value) {
  var t = document.querySelector("textarea");
  t.value = "testing";
  t.selectionStart = 1;
  t.selectionEnd = 3;
  t.selectionDirection = "backward";
  t.style.display = "";
  document.body.clientWidth;
  t.style.display = "none";
  is(t[prop], value, "Correct value for the " + prop + " property");
}

test("selectionStart", 1);
test("selectionEnd", 3);
test("selectionDirection", "backward");

</script>
</pre>
</body>
</html>
